+++ pridepy/__init__.py
@@ -1,0 +1,1 @@
+from .pridepy import main
