+++ pridepy/tests/test_authentication.py
@@ -2,8 +2,6 @@ import unittest
 
 import pytest
 
-import authentication.authentication
-
 
 class TestAuthentication(unittest.TestCase):
     """
@@ -18,7 +16,7 @@ class TestAuthentication(unittest.TestCase):
         """
         username = "******"
         password = "******"
-        auth = authentication.authentication.Authentication()
+        auth = pridepy.authentication.authentication.Authentication()
         api_token = auth.get_token(username, password)
         print(api_token)
         self.assertTrue(len(api_token) > 20, "Token not found!")
@@ -31,7 +29,7 @@ class TestAuthentication(unittest.TestCase):
         """
         username = "******"
         password = "******"
-        auth = authentication.authentication.Authentication()
+        auth = pridepy.authentication.authentication.Authentication()
         api_token = auth.get_token(username, password)
         print(api_token)
         self.assertTrue(auth.validate_token(api_token), "Token is invalid or expired!")
